third_party/blink/web_tests/css3/calc/letter-spacing-expected.txt

PASS getComputedValue(".invalid", "letterSpacing") is "20px"
PASS getComputedValue(".valid", "letterSpacing") is "10px"
PASS successfullyParsed is true

TEST COMPLETE
